Added some more tests of thing description
After the changes I made to action schemas, I wanted to ensure
I'd not broken the Thing Description.
I am no longer worried - the TD serialises
schemas in place and doesn't rely on names.  The
thing that mostly went wrong was naming the schemas,
which doesn't apply to the TD.

Also, I've tested all the cases I can think of, and the
TD still validates.
